services/market-data/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.1
kafka-python==2.0.2
pytrends==4.9.2
python-dotenv==1.0.0
pydantic==2.7.0
requests==2.31.0